backend/requirements.txt
Django==1.11
django-filter==1.0.4
djangorestframework-jwt==1.8.0
djangorestframework==3.4.3
Pillow==4.3.0
